Migrate tests to FactCheck
The main idea here is to isolate the tests from each other.
Each `facts()` or `context()` block introduces new scope.

For testing and exploration at the REPL, copy and pasting the contents
of these blocks seems to work well (but do it enough and you'll
end up with leaky scope again).