Add sub-pixel corner detection capability
Introduces the function imcorner_subpixel as well as several requisite
support functions to facilitate the ability to detect corners to
sub-pixel precision.

The imcorner_subpixel uses the existing imcorner functionality to obtain
initial corner response values and a binary corner indicator matrix. The
corner response values and the indicator matrix are used in a univariate
quadratic interpolation scheme to determine the corner coordinates to
sub-pixel precision.

The new imcorner_subpixel function returns a length-3 NTuple of type
HomogeneousPoint, which represents the homogeneous coordinates of a
planar point.