Fixes adjust_gamma so that it works on offset arrays
Calling `indices(img)` on a canonical array returns a tuple of type `Base.OneTo`. Unfortunately, there is currently no `fill` method defined in `Base` which can take a tuple of `Base.OneTo` as its dimension argument. Hence, I had to explicitly convert the tuple of `Base.OneTo` to a tuple of integers.