Adds `build_histogram` to address limitations of `imhist`
In contrast to `imhist` the `build_histogram` function returns the number of bins that the user requested.  

Furthermore, when working with an 8-bit image the `build_histogram` function will return the same bin boundaries whether the intensities are represented with a FixedPointNumber (N0f8) in the interval [0,1] or by unsigned integers in the interval [0,255]. 

The `counts` are returned as an `OffsetArray` and the zero-index of  `counts` now stores the number of intensities that were less than the first bin boundary. This constitutes a breaking change from `imhist` which did not utilise an `OffsetArray` and utilised the first index of `counts` to store the number of intensities that were less than the first bin boundary.