🧹 Move `complement` to ColorVectorSpace
`complement` is such a basic operation that we moved to lower-level package
ColorVectorSpace in PR
https://github.com/JuliaGraphics/ColorVectorSpace.jl/pull/144

This requires ColorVectorSpace v0.9.2. Since we indirectly get CVS via
ImageCore, we need to set a compatible ImageCore version that requires CVS >=
v0.9.2. For this reason, this commit sets ImageCore compatibility to v0.9.3 so
that we get CVS at least v0.9.7