🧹 Move `restrict` to ImageAxes and ImageMetadata
We used to host `restrict` in ImageTransformations, which is a big dependency
for ImageAxes and ImageMetadata. Thus specialized `restrict` methods for those
data types are kept in Images. Now that we have `restrict` in ImageBase, we can
move these specialized methods to ImageAxes and ImageMetadata.

This commit requires three new versions:

- ImageBase v0.1.2: this version fixes some of old `restrict` type inferrence issues.
- ImageAxes v0.6.9: https://github.com/JuliaImages/ImageAxes.jl/pull/59
- ImageMetadata v0.9.7: https://github.com/JuliaImages/ImageMetadata.jl/pull/79

AxisArrays is no longer needed because the method is moved to ImageAxes.

ImageTransformations <v0.9 compatibility is dropped so that we can avoid
symbol conflict on `restrict`.